TUSS4470_shield_002/web_mobile/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
numpy>=1.26.0
pyserial>=3.5
pyserial-asyncio-fast>=0.16
python-multipart>=0.0.9
